refactor(server): modulariza estructura y corrige imports en user.routes
