+++ server/.env.example
@@ -1,9 +1,7 @@
-# .env (solo en tu máquina/servidor)
-MONGO_URI=mongodb+srv://usuario:clave@cluster.xxxx.mongodb.net/kumelen
-GOOGLE_CLIENT_ID=999999999-aaaaaaa.apps.googleusercontent.com
+PORT=8080
+MONGO_URI=mongodb://localhost:27017/portal
+JWT_SECRET=changeme
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
 PERMITIDO_DOMINIO=kumelenescuela.edu.ar
-FRONTEND_URL=https://portal-kumelen.vercel.app
-JWT_SECRET=unaClaveUltraSeguraQueInventes
-JWT_EXPIRES_IN=2h
-SUPERADMINS=leandro@kumelenescuela.edu.ar,adminextra@kumelenescuela.edu.ar
-
+CORS_ORIGIN=http://localhost:5173
