+++ server/package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
+    "dev": "node --env-file=.env ./src/index.js",
+    "start": "node ./src/index.js"
   },
   "keywords": [],
   "author": "",
