+++ server/src/config/mongoose.js
@@ -1,0 +1,8 @@
+// server/src/config/mongoose.js
+import mongoose from "mongoose";
+import { env } from "./env.js";
+
+export async function connectMongo() {
+  mongoose.set("strictQuery", true);
+  await mongoose.connect(env.MONGO_URI);
+}
