+++ server/src/middlewares/error.js
@@ -1,0 +1,10 @@
+// server/src/middlewares/error.js
+export function errorHandler(err, req, res, next) {
+  const status = err.status || 500;
+  res.status(status).json({
+    error: {
+      message: err.message || "Internal Server Error",
+      code: err.code || "INTERNAL_ERROR",
+    },
+  });
+}
